lib: refactor backup functionality
Refactor backup code so it is ready for later extensions in order
to bring restore functionality.

Relates to:
https://github.com/pbrezina/authselect/issues/120
