+++ src/lib/Makefile.am
@@ -45,6 +45,7 @@ authselect_state_dir=@AUTHSELECT_STATE_DIR@
 
 libauthselect_la_SOURCES = \
     authselect.c \
+    authselect_backup.c \
     authselect_profile.c \
     authselect_files.c \
     authselect_paths.c \
